Code/Django/09_DRF/requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-extensions==3.1.3
django-seed==0.2.2
djangorestframework==3.12.4
Faker==8.1.1
freeze==0.0.1
idna==2.10
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.4
